backend/myproject/requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
attrs==22.2.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
channels==4.1.0
click==8.1.7
comm==0.1.2
constantly==23.10.4
contourpy==1.0.7
cryptography==42.0.8
cycler==0.11.0
daphne==4.1.2
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
fonttools==4.38.0
fqdn==1.5.1
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.2
jupyter_client==7.4.8
jupyter_core==5.1.3
jupyter_server==2.0.6
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.4
MouseInfo==0.1.3
mysql-connector-python==8.0.31
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.2
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
packaging==23.0
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.14.0
PyJWT==2.8.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.7.6
pyobjc-core==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-CoreText==10.1
pyobjc-framework-Quartz==10.1
pyOpenSSL==24.1.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.30
python-dateutil==2.8.2
python-json-logger==2.0.4
pytweening==1.0.7
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rubicon-objc==0.4.7
Send2Trash==1.8.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.5.0
sshtunnel==0.4.0
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.8.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uri-template==1.2.0
uvicorn==0.30.1
virtualenv==20.17.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
zope.interface==6.4.post2
